Fix ExternalSdCardOperation.getDocumentFile()
It was found that the arguments passed to String.split() in splitting given path was incorrectly interpreted, causing some SD card related operations to fail, especially when there is folder prefix(es).

For example, for a file on the SD card "tmp/DummyFile", it was interpreted as ["tmp_DummyFile"] but not ["tmp", "DummyFile"], i.e. not splitted correctly. After this change paths should be splitted correctly.