Fix Sign-in Google Button won't go away after creating Google Drive connection
Fix problem brought by #2832. Previously, after tapping the Google Drive selection, it did brought up the Cloud Plugin dialog, but did nothing afterwards, and the Sign in with Google dialog kept on screen. This changeset fixes this behaviour, by in align with behaviours of buttons of other cloud connection types.